Moved charge_extraction.py, waveform_cleaning.py and reductors.py to the images module
