+++ ctapipe/image/__init__.py
@@ -1,3 +1,6 @@
 from .hillas import *
 from .cleaning import *
 from .pixel_likelihood import *
+from .charge_extractors import *
+from .waveform_cleaning import *
+from .reductors import *
