+++ ctapipe/image/charge_extractors.py
