+++ ctapipe/image/reductors.py
